+++ CHANGES.md
@@ -9,8 +9,7 @@ CHANGES
 - Added `PERMESSAGE_DEFLATE_INVALID_MAX_WINDOW_BITS` to `WebSocketError`.
 - Added `PERMESSAGE_DEFLATE_UNSUPPORTED_PARAMETER` to `WebSocketError`.
 - Added `UNEXPECTED_RESERVED_BIT` to `WebSocketError`.
-- Modified `WebSocketFrame.toString()` to stringify payload of text and
-  binary frames.
+- Modified `WebSocketFrame.toString()` to stringify the payload better.
 - Copied Extended and AutoFlush settings in `WebSocket.recreate(int)`.
 - Added `onSendingFrame(WebSocket, WebSocketFrame)` to `WebSocketListener`.